i32 fixes and some warnings removed
Former-commit-id: 505d645e9f9d9e18b15b84625afce020e36127d8

+++ src/entity.cpp
@@ -38,9 +38,9 @@ static void DeleteEntity(game_state* GameState, u32 EntityIndex)
     GameState->EntityCount--;
 }
 
-static void DecreaseStamina(entity* Entity, game_state* GameState, i16 Cost) 
+static void DecreaseStamina(entity* Entity, game_state* GameState, i32 Cost) 
 {
-    i16 NewStamina = Max(0, Entity->Player.Stamina - Cost);
+    i32 NewStamina = Max(0, Entity->Player.Stamina - Cost);
     Entity->Player.StaminaLost = Entity->Player.Stamina - NewStamina;
     Entity->Player.Stamina = NewStamina;
     StartTimer(GameState, Entity->Player.StaminaDecreaseTimer);
